refactor: telebot listening logic for instructors ID
Previously, there was an if-else condition used to determine if the user was trying to send an input ('2021-22 Term 2, IS215, R1BW1') for the bot to query for a list of instructors OR an input ('/1', '/2', etc) to select an instructor that the bot has provided.

We are now using the .hears method with regular expressions to determine if the user is trying to select an instructor.