feat: search for course code with course name
Users are now able to search for course code with course name. A regex search is applied on each word of the term provided by the user. For example, a search term could be 'intro to program' and the search result would include Introduction to Programming.

Still need to account for searches that has results that exceed the message length limit (4096).